requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioitertools==0.12.0
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==21.1.0
arize-phoenix==4.33.2
arize-phoenix-evals==0.15.1
arize-phoenix-otel==0.5.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
backoff==2.2.1
bcrypt==3.2.0
beautifulsoup4==4.12.3
beniget==0.4.1
bleach==4.1.0
blinker==1.4
boto3==1.34.117
botocore==1.34.117
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloud-init==24.3.1
cohere==5.5.4
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.8.2
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.2.1
cramjam==2.8.3
cryptography==3.4.8
cycler==0.12.1
dataclasses-json==0.6.4
datasets==2.14.7
dbus-python==1.2.18
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
dspy-ai==2.4.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.114.2
fastavro==1.9.4
fastparquet==2024.2.0
filelock==3.13.1
FlashRank==0.2.0
flatbuffers==24.3.7
fonttools==4.51.0
frozenlist==1.4.1
fs==2.4.12
fsspec==2023.10.0
gast==0.5.2
googleapis-common-protos==1.65.0
graphql-core==3.2.4
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
hdbscan==0.8.38.post1
html5lib==1.1
httpcore==1.0.4
httplib2==0.20.2
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.21.4
humanfriendly==10.0
hyperlink==21.0.0
idna==3.6
imageio==2.34.2
imageio-ffmpeg==0.5.1
importlib_metadata==8.4.0
incremental==21.3.0
inflate64==1.0.0
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
isodate==0.6.1
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-path==0.3.2
jsonschema-spec==0.2.4
jsonschema-specifications==2023.7.1
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keyring==23.5.0
kiwisolver==1.4.5
langchain==0.1.14
langchain-community==0.0.31
langchain-core==0.1.52
langchain-openai==0.1.1
langchain-text-splitters==0.0.2
langsmith==0.1.93
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.10.0
llvmlite==0.43.0
lxml==5.1.0
lz4==3.1.3+dfsg
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.9.0
matplotlib-inline==0.1.3
more-itertools==8.10.0
motor==3.3.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
multivolumefile==0.2.3
mutagen==1.47.0
mypy-extensions==1.0.0
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.6.0
netifaces==0.11.0
notebook==6.4.8
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
onnxruntime==1.17.1
openai==1.14.2
openapi-core==0.18.2
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
openinference-instrumentation==0.1.18
openinference-instrumentation-langchain==0.1.28
openinference-instrumentation-llama-index==3.0.2
openinference-instrumentation-openai==0.1.14
openinference-semantic-conventions==0.1.10
openpyxl==3.1.2
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
optuna==3.6.0
orjson==3.10.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.0
parse==1.20.1
parso==0.8.1
pathable==0.4.3
patool==2.2.0
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.3.0
pinecone-client==3.1.0
ply==3.11
prance==23.6.21.0
proglog==0.1.10
prometheus-client==0.9.0
prompt-toolkit==3.0.28
protobuf==4.25.4
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
py==1.10.0
py7zr==0.21.0
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.1
pybars4==0.9.13
pybcj==1.0.2
pycryptodomex==3.20.0
pycurl==7.44.1
pydantic==2.9.1
pydantic_core==2.23.3
pydub==0.25.1
Pygments==2.11.2
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
PyMeta3==0.5.1
pymongo==4.6.2
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
pynndescent==0.5.13
pyodbc==5.1.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyppmd==1.1.0
pyrsistent==0.18.1
pyserial==3.5
PySocks==1.7.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.9
python-telegram-bot==21.0.1
pythran==0.10.0
pytubefix==6.5.2
pytz==2024.1
PyYAML==6.0.2
pyzmq==22.3.0
pyzstd==0.15.10
referencing==0.30.2
regex==2023.12.25
requests==2.32.3
requests-mock==1.12.1
rfc3339-validator==0.1.4
rpds-py==0.18.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.1
scikit-learn==1.4.2
scipy==1.12.0
seaborn==0.13.2
SecretStorage==3.3.1
selenium==4.18.1
semantic-kernel==0.5.0.dev0
semantic-router==0.0.65
Send2Trash==1.8.1b0
service-identity==18.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
sqlean.py==3.45.1
starlette==0.38.5
strawberry-graphql==0.236.0
sympy==1.12
systemd-python==234
tenacity==8.2.3
terminado==0.13.1
testpath==0.5.0
texttable==1.7.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.15.2
tornado==6.1
tqdm==4.66.2
traitlets==5.1.1
trio==0.25.0
trio-websocket==0.11.1
Twisted==22.1.0
types-requests==2.32.0.20240602
typesense==0.20.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
ujson==5.9.0
umap-learn==0.5.6
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==2.2.1
uvicorn==0.30.6
wadllib==1.3.6
wcwidth==0.2.5
webdriver-manager==4.0.1
webencodings==0.5.1
websockets==13.1
Werkzeug==3.0.1
widgetsnbextension==2.0.0
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
yt-dlp==2024.10.22
zipp==1.0.0
zope.interface==5.4.0
